fix(countDown): change the component's name
